[Issue-1795] Init
