+++ package.json
@@ -75,8 +75,8 @@
     "@polkadot/util": "^12.3.2",
     "@polkadot/util-crypto": "^12.3.2",
     "@subwallet/chain-list": "0.2.11-beta.0",
-    "@subwallet/keyring": "^0.0.10",
-    "@subwallet/ui-keyring": "^0.0.10",
+    "@subwallet/keyring": "^0.0.11-beta.0",
+    "@subwallet/ui-keyring": "^0.0.11-beta.0",
     "babel-core": "^7.0.0-bridge.0",
     "rxjs": "^7.8.1",
     "safe-buffer": "^5.2.1",
