+++ packages/extension-base/package.json
@@ -52,8 +52,8 @@
     "@subwallet/extension-chains": "^1.1.8-0",
     "@subwallet/extension-dapp": "^1.1.8-0",
     "@subwallet/extension-inject": "^1.1.8-0",
-    "@subwallet/keyring": "^0.0.10",
-    "@subwallet/ui-keyring": "^0.0.10",
+    "@subwallet/keyring": "^0.0.11-beta.0",
+    "@subwallet/ui-keyring": "^0.0.11-beta.0",
     "@walletconnect/sign-client": "^2.8.4",
     "@walletconnect/types": "^2.8.4",
     "@walletconnect/utils": "^2.8.4",
