+++ packages/extension-base/src/services/keyring-service/index.ts
@@ -5,6 +5,7 @@ import { CurrentAccountInfo, KeyringState } from '@subwallet/extension-base/back
 import { ALL_ACCOUNT_KEY } from '@subwallet/extension-base/constants';
 import { EventService } from '@subwallet/extension-base/services/event-service';
 import { CurrentAccountStore } from '@subwallet/extension-base/stores';
+import { InjectedAccountWithMeta } from '@subwallet/extension-inject/types';
 import { keyring } from '@subwallet/ui-keyring';
 import { SubjectInfo } from '@subwallet/ui-keyring/observable/types';
 import { BehaviorSubject } from 'rxjs';
@@ -90,6 +91,16 @@ export class KeyringService {
 
   get currentAccount (): CurrentAccountInfo {
     return this.currentAccountSubject.value;
+  }
+
+  public injectAccounts (accounts: InjectedAccountWithMeta[]) {
+    keyring.loadInjects(accounts.map((account, index) => ({
+      ...account,
+      meta: {
+        ...account.meta,
+        name: `${account.meta.name || String(index)} (${account.meta.source})`
+      }
+    })));
   }
 
   setCurrentAccount (currentAccountData: CurrentAccountInfo) {
