+++ packages/extension-koni-ui/package.json
@@ -38,9 +38,9 @@
     "@subwallet/extension-chains": "^1.1.8-0",
     "@subwallet/extension-dapp": "^1.1.8-0",
     "@subwallet/extension-inject": "^1.1.8-0",
-    "@subwallet/keyring": "^0.0.10",
+    "@subwallet/keyring": "^0.0.11-beta.0",
     "@subwallet/react-ui": "5.1.2-b69",
-    "@subwallet/ui-keyring": "^0.0.10",
+    "@subwallet/ui-keyring": "^0.0.11-beta.0",
     "@walletconnect/types": "^2.8.4",
     "boring-avatars": "^1.7.0",
     "bowser": "^2.11.0",
