+++ packages/extension-koni-ui/src/Popup/Root.tsx
@@ -52,7 +52,7 @@ function DefaultRoute ({ children }: {children: React.ReactNode}): React.ReactEl
 
   const needMigrate = useMemo(
     () => !!accounts
-      .filter((acc) => acc.address !== ALL_ACCOUNT_KEY && !acc.isExternal)
+      .filter((acc) => acc.address !== ALL_ACCOUNT_KEY && !acc.isExternal && !acc.isInjected)
       .filter((acc) => !acc.isMasterPassword)
       .length
     , [accounts]
